feat: add intent filters for tasker automation (#119)
