+++ common/src/main/java/com/github/kr328/clash/common/constants/Intents.kt
@@ -5,6 +5,9 @@ import com.github.kr328.clash.common.util.packageName
 object Intents {
     // Public
     val ACTION_PROVIDE_URL = "$packageName.action.PROVIDE_URL"
+    val ACTION_START_CLASH = "$packageName.action.START_CLASH"
+    val ACTION_STOP_CLASH = "$packageName.action.STOP_CLASH"
+    val ACTION_TOGGLE_CLASH = "$packageName.action.TOGGLE_CLASH"
 
     const val EXTRA_NAME = "name"
