+++ design/src/main/res/values-zh/strings.xml
@@ -245,4 +245,7 @@
     <string name="geofile_imported">%1$s 已导入</string>
     <string name="toast_profile_updated_complete">更新配置 %s 成功</string>
     <string name="toast_profile_updated_failed">更新配置 %1$s 失败: %2$s</string>
+    <string name="external_control_activity">External Control</string>
+    <string name="external_control_started">Clash.Meta 服务已启动</string>
+    <string name="external_control_stopped">Clash.Meta 服务已停止</string>
 </resources>
