+++ design/src/main/res/values/strings.xml
@@ -310,4 +310,7 @@
     <string name="geofile_imported">%1$s imported</string>
     <string name="toast_profile_updated_complete">Update profile %s completed</string>
     <string name="toast_profile_updated_failed">Update profile %1$s failed: %2$s</string>
+    <string name="external_control_activity">External Control</string>
+    <string name="external_control_started">Clash.Meta service started</string>
+    <string name="external_control_stopped">Clash.Meta service stopped</string>
 </resources>
